pyproject.toml
[tool.poetry]
name = "airfrance-klm-api"
version = "0.1.0"
description = "This Python package provides a simple and easy-to-use interface for interacting with Air France's Developer API."
authors = ["Zach Bloss <zacharybloss@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "airfrance_klm_api"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
pydantic = "^2.3.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
isort = "^5.12.0"
black = "^23.7.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
